requirements.txt
aiotube==1.7.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
dataclasses-json==0.5.14
ftfy==6.1.1
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
marshmallow==3.20.1
mypy-extensions==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-youtube==0.9.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
youtube-scraping-api==1.1.4
